fix: 엔티티 Id 데이터 타입 Integer로 변경

+++ src/main/java/org/vuetiful/DNS/domain/notification/entity/Notification.java
@@ -14,7 +14,7 @@ public class Notification {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long notificationId;
+    private Integer notificationId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "memberId", nullable = false)
