+++ src/main/java/org/vuetiful/DNS/domain/postImage/entity/PostImage.java
@@ -14,7 +14,7 @@ public class PostImage {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long postImageId;
+    private Integer postImageId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "postId", nullable = false)
